fixed bugs with db

+++ Passtore/src/main/java/util/SQLiteConnector.java
@@ -3,10 +3,23 @@ package util;
 import java.sql.*;
 import model.*;
 
+/**
+ * Class used to connect to the database.
+ */
 public class SQLiteConnector{
     private static Connection conn;
+    
+    /**
+     * Directory of the database saved.
+     */
     private static String url;
 
+    /**
+     * Connects to the database.
+     * Location of the database depending on which OS it is.
+     * if linux, stored in home/user/
+     * if windows, store in C:/Users/username/
+     */
     public static void connect(){
         if (System.getProperty("os.name").equals("Linux")) url = "jdbc:sqlite:" + System.getProperty("user.home") + "/passtoresavedata.db";
         else url = "jdbc:sqlite:C:\\Users\\" + System.getProperty("user.name") + "\\passtoresavedata.db";
@@ -21,16 +34,27 @@ public class SQLiteConnector{
         }
     }
 
+    /**
+     * Creates the tables if they don't exist
+     * Called from connect()
+     * @throws SQLException
+     */
     public static void createTables() throws SQLException{
         Statement s = conn.createStatement();
 
         String line = "CREATE TABLE IF NOT EXISTS masteraccounts(username TEXT PRIMARY KEY, password TEXT );";
         s.executeUpdate(line);
 
-        line = "CREATE TABLE IF NOT EXISTS accounts( owner TEXT, site TEXT PRIMARY KEY, password TEXT, email TEXT, username TEXT );";
+        line = "CREATE TABLE IF NOT EXISTS accounts( owner TEXT , site TEXT , password TEXT, email TEXT, username TEXT," +
+                                                        " PRIMARY KEY (owner, site) );";
         s.executeUpdate(line);
     }
 
+    /**
+     * Used for running queries
+     * @param query SELECT statement to run
+     * @return the result set received from the query
+     */
     private static ResultSet getResultSet(String query){
         ResultSet rs = null;
         try{
@@ -43,6 +67,10 @@ public class SQLiteConnector{
         return rs;
     }
 
+    /**
+     * Used for insert, update and delete statements
+     * @param line statement to run on the database
+     */
     public static void update(String line){
         try{
             Statement s = conn.createStatement();
@@ -52,16 +80,18 @@ public class SQLiteConnector{
         }
     }
 
-    public static void updateMasterAccount(String oldUsername, MasterAccount updated){
-        update("UPDATE masteraccounts SET username = '" + updated.getUsername() + "'," +
-                                            "password = '" + updated.getPassword() + "' WHERE username = '" + oldUsername + "'");
+    public static void updateMasterAccount(String oldUsername, MasterAccount ma){
+        update("UPDATE accounts SET owner = '" + ma.getUsername() + "' WHERE owner = '" + oldUsername + "'");
+        update("UPDATE masteraccounts SET username = '" + ma.getUsername() + "'," +
+                                            "password = '" + ma.getPassword() + "' WHERE username = '" + oldUsername + "'");
     }
 
-    public static void updateAccount(String oldSite, Account updated){
-        update("UPDATE accounts SET username = '" + updated.getUsername() + "'," +
-                                            "password = '" + updated.getPassword() + "'," +
-                                            "site = '" + updated.getSite() + "'," + 
-                                            "email = '" + updated.getEmail() + "' WHERE site = '" + oldSite + "'");
+    public static void updateAccount(String oldSite, Account a, MasterAccount ma){
+        update("UPDATE accounts SET username = '" + a.getUsername() + "'," +
+                                            "password = '" + a.getPassword() + "'," +
+                                            "site = '" + a.getSite() + "'," + 
+                                            "email = '" + a.getEmail() + "' WHERE site = '" + oldSite + "'"
+                                            + " AND owner = '" + ma.getUsername() + "'");
     }
 
     public static void addMasterAccount(MasterAccount ma){
@@ -76,18 +106,19 @@ public class SQLiteConnector{
 
     public static void deleteMasterAccount(MasterAccount ma){
         update("DELETE FROM masteraccounts WHERE username = '" + ma.getUsername() + "'");
+        update("DELETE FROM accounts WHERE owner = '" + ma.getUsername() + "'");
     }
 
-    public static void deleteAccount(Account a){
-        update("DELETE FROM accounts WHERE site = '" + a.getSite() + "'");
+    public static void deleteAccount(Account a, MasterAccount ma){
+        update("DELETE FROM accounts WHERE site = '" + a.getSite() + "' AND owner = '" + ma.getUsername() + "'");
     }
 
     public static ResultSet getMasterAccounts(){
         return getResultSet("SELECT * FROM masteraccounts");
     }
 
-    public static ResultSet getAccounts(String masterUsername){
-        return getResultSet("SELECT * FROM accounts WHERE owner = '" + masterUsername + "'");
+    public static ResultSet getAccounts(MasterAccount ma){
+        return getResultSet("SELECT * FROM accounts WHERE owner = '" + ma.getUsername() + "'");
     }
 
 }
